# Add support of Parallel GET/PUT for tokio