… s3/src/utils.rs