@durch 2022-09-24
@kerryeon I'm gonna close this, I've added parallel support for async uploads using `futures::future::join_all`, thanks for inspiring me to do it, my guess is that the performance is a bit worse then what you did, but the implementation is a bit simpler :)

It would be great if you could test out the new `PUT` implementation, I didn't touch `GET` yet.

[Relevant change](https://github.com/durch/rust-s3/commit/6d8b6322997eda96c70ab3b5099b28e811516f6a#diff-fc8c41c08a07d1c4a7bcaa60e706309b5aa34a740b37ca16e9779dcc477d00efR1033)