@kerryeon 2022-10-27
**Environments**: 
* NIC: Mellanox ConnectX-5 100Gb
* S3: Rook-Ceph Object Storage (Bare-metal)

**Experiments**: 
* Data Size: 64 MiB
* Number of Iterations: 30
* Read Speed (NO parallel): **11.14 Gbps**
* Read Speed (Parallel): 10.30 Gbps
* Write Speed (NO multipart): 337.07 Mbps
* Write Speed (Proposed Parallel): **1.01 Gbps**
* Write Speed (Current): 324.66 Mbps

This implementation has no profits of using multipart uploading.

By the way, you said you were inspired by my implementation, but I'm embarrassed that this statement is not quoted anywhere in commit, release, etc. It gives me the impression that my contribution has really helped nowhere.